Desktop/port_back/requirements.txt
asgiref==3.2.10
Django==3.1.1
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
python-dotenv==0.14.0
django-cors-headers==3.5.0
django-pgcrypto-fields==2.5.2
psycopg2==2.8.6
drf-yasg==1.17.1
PyQRCode==1.2.1
pypng==0.0.20
django-otp==1.0.0
python-u2flib-server==5.0.0
drf-nested-routers==0.91
django-filter==2.4.0
celery==4.4.6
redis
requests~=2.24.0
xhtml2pdf~=0.2.5
python-dateutil~=2.8.1
imgkit~=1.0.2
Pillow~=8.0.0
PyPDF2~=1.26.0
reportlab~=3.5.53
sentry-sdk==0.19.1
pycrypto==2.6.1
docxtpl==0.11.1
django-ses
django-prometheus
pycryptodomex
python-logstash